Fix zathura not accepting stdin due to '&' sign
It seems that the trailing '&' sign after the zathura command makes
zathura open before it can finish reading the standard input, causing it
to fail when the '-' option is used.

The '&' doesn't need to be inside the script - if you want it, you can
just add it yourself on the command line.